# feat: use `ArrayPool.Rent` to prevent `byte[]` allocation