--- a/QRCoder/PngByteQRCode.cs
+++ b/QRCoder/PngByteQRCode.cs
@@ -97,12 +97,12 @@ public sealed class PngByteQRCode : AbstractQRCode, IDisposable
         var quietZoneOffset = (drawQuietZones ? 0 : 4);
         var bytesPerScanline = (matrixSize * pixelsPerModule + 7) / 8 + 1; // A monochrome scanline is one byte for filter type then one bit per pixel.
         var scanLinesLength = bytesPerScanline * matrixSize * pixelsPerModule;
-        #if NETCOREAPP2_1_OR_GREATER || NETSTANDARD2_1
+#if NETCOREAPP2_1_OR_GREATER || NETSTANDARD2_1
         var scanlines = ArrayPool<byte>.Shared.Rent(scanLinesLength);
         Array.Clear(scanlines, 0, scanLinesLength);
-        #else
+#else
         var scanlines = new byte[scanLinesLength];
-        #endif
+#endif
 
         for (var y = 0; y < matrixSize; y++)
         {