WIP need to do benchmarks and handle the array length exceeding `scanLinesLength`