@TimothyMakkison 2025-10-01
Okay, did my TODO list, I've used `ArraySegment<byte>` to make it harder for a longer rented array to break things 🤞 Looks like its has decent memory savings.


Have you considered using a formatter like csharpier for this project? Iirc Mapperly has it to automatically format code as its pushed, would save code rabbit from rescanning the commit

### Benchmarks
#### Before
| Method              | Mean        | Error     | StdDev    | Gen0   | Allocated |
|-------------------- |------------:|----------:|----------:|-------:|----------:|
| CreateQRCode        |    150.3 us |   2.93 us |   2.74 us | 0.7324 |   7.14 KB |
| CreateQRCodeLong    |  2,484.3 us |  32.43 us |  25.32 us |      - |  33.18 KB |
| CreateQRCodeLongest | 14,196.0 us | 217.65 us | 169.93 us |      - |  79.63 KB |
#### After (different day, timings are not reliable)
| Method              | Mean        | Error     | StdDev    | Gen0   | Allocated |
|-------------------- |------------:|----------:|----------:|-------:|----------:|
| CreateQRCode        |    93.90 us |  0.682 us |  0.605 us | 0.3662 |   4.07 KB |
| CreateQRCodeLong    | 1,747.93 us | 16.420 us | 15.360 us |      - |  11.05 KB |
| CreateQRCodeLongest | 9,770.56 us | 58.106 us | 48.521 us |      - |  46.42 KB |