@Shane32 2025-10-01
I don't care too much about Coderabbit's checks; I noticed it was reading the build check outputs here and providing a suggestion based on that.  Coderabbit is nice sometimes but can also be a bit verbose (and sometimes wrong).  In PR #611 just now it correctly caught a bug unrelated to the PR :+1: .  And it's always nice that it writes a PR summary.