@TimothyMakkison 2025-10-01
Reopening, I realised that my benchmarks were wrong and that they weren't measuring a PNG qr code. Here are the actual results.

### Before
<img width="1616" height="476" alt="image" src="https://github.com/user-attachments/assets/2ab84d41-be24-4ffb-ba44-6a93439acb36" />


### After
<img width="1594" height="422" alt="image" src="https://github.com/user-attachments/assets/2305cd79-97ab-4b4b-8c91-48624f7eaa8f" />